+++ src/components/new/view/step1/BasicStep.tsx
@@ -1,12 +1,12 @@
 import { Book } from "@/apis/book/vo/book";
 import { ReadingStatus } from "@/apis/review/vo/readingStatus";
-import { ErrorMessage } from "@/components/common/ErrorMessage";
+import { ErrorMessage } from "@/components/common/form/ErrorMessage";
 import { FunnelStepComponentProps } from "@/hooks/funnel/types";
 import { isNil, isNotNil } from "@/utils/TypeUtil";
 import { zodResolver } from "@hookform/resolvers/zod";
 import { useForm } from "react-hook-form";
-import { FormFrame } from "../../../common/FormFrame";
-import { FormItem } from "../../../common/FormItem";
+import { FormFrame } from "../../../common/form/FormFrame";
+import { FormItem } from "../../../common/form/FormItem";
 import { BookReviewFunnelContext, createBasicStepSchema } from "../../types/stepContext";
 
 interface Props extends FunnelStepComponentProps<BookReviewFunnelContext, "step1"> {
